tc: add a semantics analyzer
Semantic analysis in a compiler include type checking, name resolution,
and scope analysis.

Type checking involves verifying that the types of expressions and variables
are consistent with the rules of the language.

Name resolution involves determining the meaning of identifiers (such as
variables and functions) in the context of the program.

Scope analysis involves determining the visibility and accessibility of
identifiers in different parts of the program.

- use hash list to implement symbol table
- use symbol table to implement scope

Signed-off-by: Liu Yuan <namei.unix@gmail.com>