tc: add basic generator
We just generate x86 asm code directly as an IR and call gcc to assemble
it directly.

For now we just support simple return expression statement.

$./tc -s 'int main(){ return 1+2*3-4;}' # generate an executable as a.tc
$./a.tc;echo $? # test exit code of the a.tc

we use stack for expression evaluation. THIS IS NOT EFFETIENT AT ALL but
it is simple to understand.

Someday we might revisit it and optimize it as cool as real compiler
like gcc.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>